requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2.10
djangorestframework==3.13.1
dnspython==2.1.0
gunicorn==20.1.0
idna==2.10
Naked==0.1.31
psycopg2==2.9.3
pycryptodome==3.10.1
pymongo==4.0.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.5
uuid==1.30
whitenoise==6.1.0
